requirements.txt
functions-framework
google-cloud-storage
requests
beautifulsoup4
openai
pytest